pyproject.toml
[project]
name = "eventbooking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.8.1",
    "load-dotenv>=0.1.0",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
]
